feat: updating config tables and add fetching data route
